requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
PyMySQL==1.1.0
WTForms==3.1.1
authlib==1.3.0
python-dotenv==1.0.0
bcrypt==4.1.2
Gunicorn==21.2.0
eventlet==0.35.1
requests==2.31.0
Pillow==10.2.0
Markdown==3.5.2
